pyproject.toml
[project]
name = "csda-client"
version = "0.1.0"
description = "API client for CSDA services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["httpx>=0.28.1", "pydantic>=2.11.7", "stapi-pydantic>=0.0.4"]

[project.urls]
repository = "https://github.com/NASA-IMPACT/csda-client"

[dependency-groups]
dev = ["mypy>=1.15.0", "pytest>=8.3.5", "ruff>=0.11.9", "pytest-dotenv>=0.5.2"]
docs = [
    "contextily>=1.6.2",
    "geopandas>=1.0.1",
    "ipykernel>=6.29.5",
    "pystac-client>=0.8.6",
    "humanize>=4.12.3",
    "tabulate>=0.9.0",
    "cql2>=0.3.7",
    "jinja2>=3.1.6",
    "basemap>=2.0.0",
    "rasterio>=1.4.3",
    "contextily>=1.6.2",
    "json-schema-for-humans>=1.4.1",
]

[tool.uv]
default-groups = ["dev", "docs"]

[tool.mypy]
files = "src/**/*.py,tests/**/*.py"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
